stage1-module5-task1/src/main/java/com/epam/mjc/InterfaceCreator.java
package com.epam.mjc;

import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.function.*;
import java.util.stream.Collectors;

public class InterfaceCreator {

    public Predicate<List<String>> isValuesStartWithUpperCase() {
        return x -> x.stream().allMatch(s -> Character.isUpperCase(s.charAt(0)));
    }

    public Consumer<List<Integer>> addEvenValuesAtTheEnd() {
        return x -> x.stream().filter(n -> n % 2 == 0)
                .collect(Collectors.toList())
                .forEach(y -> x.add(y));
    }

    public Supplier<List<String>> filterCollection(List<String> values) {
        return () -> values.stream()
                .filter(x -> Character.isUpperCase(x.charAt(0)))
                .filter(x -> x.endsWith("."))
                .filter(x -> x.split(" ").length > 3)
                .collect(Collectors.toList());
    }

    public Function<List<String>, Map<String, Integer>> stringSize() {
        return x -> {
            Map<String, Integer> map = new HashMap<>();
            for (String str : x) {
                map.put(str, str.length());
            }
            return map;
        };
    }

    public BiFunction<List<Integer>, List<Integer>, List<Integer>> concatList() {
        return (list1, list2) -> {
            for (int x : list2) {
                list1.add(x);
            }
            return list1;
        };
    }
}
